Improvement error handling. Refactoring.
- fix error 'LDAP user not found in local DB' handling
- fix error handling in WebController: return viewname, not message
