+++ project-TaskManagementSystem/src/main/java/ru/otus/WebSecurityConfig.java
@@ -9,6 +9,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import ru.otus.auth.CustomUserDetailsContextMapper;
 
 @Configuration
 @EnableWebSecurity
