+++ project-TaskManagementSystem/src/main/java/ru/otus/auth/UserNotFoundInLocalDbException.java
@@ -1,0 +1,13 @@
+package ru.otus.auth;
+
+import org.springframework.security.authentication.AccountStatusException;
+
+public class UserNotFoundInLocalDbException extends AccountStatusException {
+    public UserNotFoundInLocalDbException(String msg) {
+        super(msg);
+    }
+
+    public UserNotFoundInLocalDbException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+}
