+++ project-TaskManagementSystem/src/main/java/ru/otus/service/AuthenticatedUserInfoService.java
@@ -3,6 +3,7 @@ package ru.otus.service;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.ldap.userdetails.LdapUserDetailsImpl;
 import org.springframework.stereotype.Service;
+import ru.otus.auth.UserNotFoundInLocalDbException;
 import ru.otus.model.User;
 
 @Service
@@ -14,10 +15,13 @@ public class AuthenticatedUserInfoService {
         this.userService = userService;
     }
 
-    public User getAuthenticatedUser(){
+    public User getAuthenticatedUser() throws IllegalStateException {
         var ldapUserDetails = (LdapUserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         String username = ldapUserDetails.getUsername();
         var optionalUser = userService.findByUsername(username);
-        return optionalUser.orElse(null);
+        if(optionalUser.isPresent()) {
+            return optionalUser.get();
+        }
+        throw new UserNotFoundInLocalDbException("User \"" + username + "\" not found in local database");
     }
 }
